Conflicts:
	services/api/test/fixtures/links.yml

+++ services/api/lib/has_uuid.rb
@@ -36,7 +36,9 @@ module HasUuid
   end
 
   def destroy_permission_links
-    Link.destroy_all(['link_class=? and (head_uuid=? or tail_uuid=?)',
-                      'permission', uuid, uuid])
+    if uuid
+      Link.destroy_all(['link_class=? and (head_uuid=? or tail_uuid=?)',
+                        'permission', uuid, uuid])
+    end
   end
 end
